requirement.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain_text_splitters
langchain_openai
Chromadb
sentence_transformers
langchain_huggingface
huggingface_hub[hf_xet]
faiss-cpu 
langchain_chroma
langchain_groq
langchain_xai
streamlit
langchain_core
fastapi
uvicorn
langserve
sse_starlette